refactor(dashboards): StudentDashboard sous 300 lignes — éclatement complet (H10)
549 → 152 lignes. Logique (cache + rafraîchissement arrière-plan) extraite en composable useStudentDashboard (console.log/appels KLASSCI conservés). Skeleton de chargement → DashboardSkeleton ; contenu (profil/widgets/notes/actions) → DashboardWidgets (prop dashboardData, router-links identiques). CSS verbatim. Parité stricte.

Dette assumée : .widget-card/.widget-header dupliqués verbatim entre DashboardSkeleton et DashboardWidgets (composants scoped indépendants). DashboardWidgets : commentaires de section retirés pour tenir ≤300 (aucune déclaration CSS modifiée).
